generate a module for definitions nested inside records
In C++, a class body introduces a scope which can contain typedefs,
enums, and other record definitions. In Rust, structs can have
associated constants but not associated type definitions or aliases, so
if a class in the C++ source holds nested definitions, we introduce a
module called "ClassName_" in the generated Rust to hold those
definitions. We also place the vtable and any unnamed record types
inside this module.