perform array-to-pointer decay by forming a pointer to the array type rather than to the element type
Previously we performed array-to-pointer decay by forming a pointer to
the element type of the array type. For statically sized arrays
(currently the only type of arrays we handle), instead form a pointer to
the statically sized array, which allows us to preserve more type
information in the generated Rust bindings.