add IUnknown methods to Interface trait
ComPtr is defined in the com-scrape-types crate and FUnknown is
generated as part of the vst3-bindgen crate, but ComPtr needs to be able
to call FUnknown methods for reference counting and interface casting.
Add query_interface, add_ref, and release methods to the Interface trait
so that ComPtr can access them without having to know about FUnknown
specifically.

For now, this is done by generating an invocation of an impl_interface!
macro for each generated interface definition and then including the
actual definition of impl_interface! in vst3-bindgen's lib.rs. There may
be a cleaner way to do this.