add ComRef and implement reference counting for ComPtr
Add ComRef<'a, I>, which represents a non-owning reference to a COM
object and does not interact with the object's reference count. Have
ComPtr increment an object's reference count in its Clone impl and
decrement it in its Drop impl. Add a ComRef::to_com_ptr() method which
also increments the object's reference count and a ComPtr::as_com_ref()
method which returns a ComRef with a lifetime constrained by the
corresponding ComPtr. Add tests for the reference counting
functionality.