cross-compilation support
Determine whether we are cross-compiling by comparing the values of the
TARGET env var during the execution of the com-scrape build script and
the execution of com-scrape itself (during the vst3-bindgen build
script). If the values are different, pass the corresponding target
triple as an argument to libclang to ensure that we generate bindings
for the target platform rather than the host platform.