fix bug where Interface::inherits() impls were not checking against base classes
The check for whether to include a recursive call to
Interface::inherits() for an interface's base class was incorrectly
always returning false, since it was checking the "bases" iteration
variable from a previous section of the method. Change it to correctly
check against the original "record.bases", and tighten up the scope for
the base class traversal loops since that would have prevented this
mistake from occurring.