Use runtime clang-sys feature to match cpal
I was getting the following `clang-sys` failure when using `vst3-bindgen` in a
crate that also depends on `cpal`:

```
  thread 'main' panicked at 'a `libclang` shared library is not loaded on this thread', /Users/kevin/.cargo/registry/src/index.crates.io-6f17d22bba15001f/clang-sys-1.6.1/src/lib.rs:1735:1
```

I think `cpal`'s transitive dependency on `clang-sys` enables the runtime
feature, also enabling it for `com-scrape`'s usage of `clang-sys`. But
`clang_sys::load()` must be called when the runtime feature is enabled.