always pass -target flag to clang
Currently, com-scrape only passes the `-target` flag to clang if the `TARGET`
env var is present and it doesn't match the value of `HOST_TARGET` (which is
set to the `TARGET` env var passed to com-scrape's build script). Instead,
always pass the value of `TARGET` if it is present, and `HOST_TARGET`
otherwise.