add Send and Sync impls for ComPtr and ComRef (#14)
ComPtr<T> and ComRef<T> currently do not implement Send or Sync, regardless of what T is, since they contain NonNull fields. This is unnecessarily restrictive. Add Send and Sync impls for ComPtr and ComRef with the same bounds as the Send and Sync impls for Arc in the standard library.