skip result values in generator and provide manual definitions
The result values in funknown.h (such as `kNoInterface`) have different values
on Windows vs. on other platforms, and libclang only gives us the set of values
for the current platform, which prevents us from generating a single set of
bindings which will be valid for all platforms. To work around this, skip the
result values in the generator and provide manual Rust definitions instead.