default to int on all platforms if an enum does not specify a base type
When an enum declaration does not specify an explicit base type, Clang defaults
to different integer types on different platforms (`int` vs. `unsigned int`).
However, we want to make a consistent choice across all platforms. Libclang
does not provide an explicit API for detecting this situation (i.e., on a
platform where Clang defaults to using `int`, there is no way to distinguish
between an enum with a base type of `int` and an enum without a base type), so
we determine it by parsing the beginning of the enum declaration ourselves.

The parser implemented in this commit will not handle every possible situation,
but it should at least produce a clear error for unexpected inputs, and it can
be extended to handle additional situations in the future.