override result value constants so that generator output is consistent across platforms
The result value constants in the VST 3 headers (`kNoInterface`, `kResultOk`,
etc.) have different values on Windows vs. on other platforms, but we want to
output the same generated bindings for all platforms. Add an
`override_constant_values` option to `Generator`, and use it to override the
result value constants in the generated output to refer to a set of definitions
in the `support` module (which in turn use `#[cfg]` to take different values on
Windows vs. other platforms).