override types of enums without explicit base type so that generator output is consistent across platforms
Enums that do not specify an explicit base type default to different underlying
types on different platforms (`int` on Windows and `unsigned int` elsewhere).
Ideally, we would be able to detect this case in the parser and handle it
automatically, but libclang does not provide any reliable interface for this
(an enum without a base type, defaulting to `int`, appears exactly the same as
an enum with an explicit base type of `int`). Since we want to output the same
generated bindings on all platforms, we are forced to override the types of
these enum declarations manually (using an `override_typedef_types` option on
the generator).

Additionally, since the `PhysicalUITypeIDs` enum is signed on some platforms
and unsigned on others, the generator outputs a different integer literal for
the value of the `kInvalidPUITypeID` constant, so we provide a manual override
for its value as well.